Add functions for sending PUT, POST and DELETE with timeout
Add oc_do_post_with_timeout, oc_do_put_with_timeout and
oc_do_delete_with_timeout functions, which take timeout argument.
When the timeout is reached the POST/PUT/DELETE handler gets
invoked with OC_STATUS_SERVICE_UNAVAILABLE status code.

+++ api/oc_introspection.c
@@ -86,7 +86,7 @@ oc_core_introspection_data_handler(oc_request_t *request,
   } else {
     OC_ERR(
       "oc_core_introspection_data_handler : %ld is too big for buffer %ld \n",
-      IDD_size, OC_MAX_APP_DATA_SIZE);
+      IDD_size, (long)OC_MAX_APP_DATA_SIZE);
     request->response->response_buffer->response_length = 0;
     request->response->response_buffer->code =
       oc_status_code(OC_STATUS_INTERNAL_SERVER_ERROR);
